Add a check that the returned list of files from the FC is not empty, otherwise return S_OK()
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@33538 4525493e-7705-40b1-a816-d608a930855b

+++ TransformationSystem/Agent/TransformationCleaningAgent.py
@@ -309,6 +309,9 @@ class TransformationCleaningAgent( AgentModule ):
     if not res['OK']:
       return res
     fileToRemove = res['Value']
+    if not len(fileToRemove):
+      gLogger.info('No files found for transID %s'%transID)
+      return S_OK()
     res = self.replicaManager.removeFile( fileToRemove )
     if not res['OK']:
       return res
